fix (refs T33506): Move remaining CustomFactory-Party into CustomFactory folder
